% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_mm_features.R
\name{amend_mm_features}
\alias{amend_mm_features}
\title{Create final features out of the variables}
\usage{
amend_mm_features(df, mm_source = "self_report", verbose = FALSE)
}
\arguments{
\item{df}{The input data frame.}

\item{mm_source}{The source of disease history.}

\item{verbose}{Whether to print progress to console.}
}
\value{
The input data frame with consolidated disease indicator columns
and the intermediate \verb{_add_N} columns removed.
}
\description{
\code{amend_mm_features} takes the data frame and amends the MM features with
additional data.
}
