% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_balance.R
\name{correct_balance}
\alias{correct_balance}
\title{Correct class imbalance}
\usage{
correct_balance(
  df,
  target_var,
  remove_vars = NULL,
  approach = NULL,
  balance_prop = 0.5,
  random_seed,
  verbose = FALSE,
  ordinals = c(),
  reverse_dummies = TRUE,
  K = 5,
  normalise = FALSE
)
}
\arguments{
\item{df}{The input data frame.}

\item{target_var}{The target variable for which the class imbalance should be
corrected.}

\item{remove_vars}{A list of column names that should be removed prior to
balance correction. Especially useful for SMOTE.}

\item{approach}{The approach by which class imbalance should be corrected.
The possible choices are 'downsample' and 'SMOTE'.}

\item{balance_prop}{The desired proportion of the total for the
current minority class. 0.5 will completely balance the classes.}

\item{random_seed}{The random seed run at the start of the function.}

\item{verbose}{Whether to print progress to console.}

\item{ordinals}{Relevant for SMOTE. A list of column names indicating
the ordinal variables in the data frame. All character/factor columns
not included in this list will be considered as nominal variables, and
all numerical variables as numerical.}

\item{reverse_dummies}{Relevant for SMOTE. Whether dummy variables created for
nominal categorical variables should be turned back into their original form.
This reduces the number of variables so that e.g., we have "sex" instead of
"sex.0" and "sex.1".}

\item{K}{Relevant for SMOTE. The number of nearest neighbours for KNN.}

\item{normalise}{Whether numerical variables should be scaled. Defaults to TRUE
when \code{approach = 'SMOTE'}}
}
\value{
The input data frame after class-imbalance correction. When
\code{approach = NULL}, the data frame is returned unchanged (except for
potential column removal and scaling).
}
\description{
\code{correct_balance} takes a data frame and corrects class imbalance for a
chosen binary variable. It returns the balanced data frame.
}
