% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_indicator.R
\name{disease_indicator}
\alias{disease_indicator}
\title{Create a disease indicator from disease date}
\usage{
disease_indicator(
  df,
  colname,
  date_form = NULL,
  baseline_date = NULL,
  invalid_dates = NULL,
  disease_time = FALSE,
  time_units = "days",
  threshold_date = NULL
)
}
\arguments{
\item{df}{Input data frame.}

\item{colname}{Character; Name of the column with the dates for the event.}

\item{date_form}{Character; format of the date. Ignored for arguments that
are already of class \code{Date}.}

\item{baseline_date}{String indicating the baseline date. The string should
match a variable in \code{df} that contains character strings or Date values
convertible to dates.}

\item{invalid_dates}{Character or Date vector; invalid dates in the form
\code{date_form} (or already of class \code{Date}). Events with these dates will
be converted to NA.}

\item{disease_time}{Logical; whether the time between the baseline
date and the disease date should be calculated and added to the data frame.
Units are controlled by \code{time_units}.}

\item{time_units}{Character; unit for the disease time column. One of
\code{'days'}, \code{'weeks'}, \code{'months'}, \code{'years'}. Default is \code{'days'}.}

\item{threshold_date}{Single character/Date or vector of same length as
\code{nrow(df)}; should be in the same format as \code{colname} (or already of class
\code{Date}). If not \code{NULL}, events occurring after this date are considered
non-events and changed to NA in the returned data frame.}
}
\value{
The input data frame with the original event column renamed to
\verb{<colname>_date} and a new integer indicator column \verb{<colname>} (1 = event,
0 = no event, \code{NA} = invalid date).
}
\description{
\code{disease_indicator} takes a data frame and a string indicating the column
containing the dates of an event, and returns the same data frame with
a new indicator column for that event. The original event column names is
changed; the new column has the same name as the original column for the
event.
}
