% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_classify.R
\name{education_classify}
\alias{education_classify}
\title{Classify education}
\usage{
education_classify(x)
}
\arguments{
\item{x}{The input array.}
}
\description{
\code{education_classify} takes an array of coded qualifications answers
according to UKB data field 6138 and classifies it into three classes.
The higher activity level is prioritised in cases of answers fitting into
multiple categories per participant.
}
