% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit_extract.R
\name{matchit_extract}
\alias{matchit_extract}
\title{Extract simplified table of matching metrics within \code{matchit_comparison}.}
\usage{
matchit_extract(bal, imputed, smd_thresh = 0.05)
}
\arguments{
\item{bal}{The input list; should be the 'results' object in the output list
of \code{matchit_comparison}.}

\item{smd_thresh}{The SMD threshold for what is considered a good match.}

\item{impute}{Whether the matching was performed on an imputed dataset.}
}
\value{
A one-row tibble with balance metrics including max/mean SMD,
number of variables exceeding the SMD threshold, worst/mean variance ratio,
and counts of discarded observations, or \code{NULL} if the input is \code{NULL}.
}
\description{
\code{matchit_extract} takes as input the 'results' object of the \code{matchit_comparison}
function to create a data frame with simplified matching metrics,
with matching approaches as rows and metrics as columns.
}
