% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phys_act_classify.R
\name{phys_act_classify}
\alias{phys_act_classify}
\title{Classify physical activity}
\usage{
phys_act_classify(x)
}
\arguments{
\item{x}{The input array.

0 - none, low light household tasks only
1 - medium heavy household tasks and/or walking for pleasure and/or other exercise,
2 - high strenuous sports.}
}
\description{
\code{phys_act_classify} takes an array of coded physical activity answers
according to UKB data field 6164 and classifies it into "none", "mild",
"moderate", or "heavy" physical activity. The higher activity level is
prioritised in cases of answers fitting into multiple categories per
participant.
}
