% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mm.R
\name{prep_mm}
\alias{prep_mm}
\title{Combines training preparation and training itself into a single function. To
be used with multimorbidity predictors.}
\usage{
prep_mm(
  input_file_path = NULL,
  output_file_path = NULL,
  extra_suffix,
  mm_source,
  target_var,
  train_features,
  max_followup,
  remove_censored = FALSE,
  min_age,
  imbalance_correct,
  remove_vars = NULL,
  balance_prop,
  ordinals,
  smote_K,
  verbose,
  random_seed,
  prep = TRUE,
  train = TRUE
)
}
\arguments{
\item{input_file_path}{The directory name from which to load the input file.}

\item{output_file_path}{The directory name into which the output file
is to be saved.}

\item{extra_suffix}{An optional extra suffix to be added to the output file;
by default, the MM source and target variable are already in the name to
prevent accidental overwrites during export.}

\item{mm_source}{The source of disease history.}

\item{target_var}{The target variable to be predicted.}

\item{max_followup}{The time within which the outcome that is to be
predicted should occur. If not set to \code{NULL}, the input data frame must
contain a column called 'followup' which is the time in years between baseline
and right-censoring or event occurrence, whichever occurs first.}

\item{remove_censored}{Whether participants that were lost to follow-up
before max_followup should be removed. Assumes the presence of the columns
\code{asc_date} and \code{cutoff_date} in the data frame.}

\item{min_age}{The minimum age of the sample. Determines whether participants
below a certain age should be dropped prior to training.}

\item{imbalance_correct}{The desired proportion of the total for the
current minority class. 0.5 will completely balance the classes.}

\item{remove_vars}{A list of column names that should be removed prior to
balance correction. Especially useful for SMOTE.}

\item{balance_prop}{The desired proportion of the total for the
current minority class. 0.5 will completely balance the classes.}

\item{ordinals}{Relevant for SMOTE. A list of column names indicating
the ordinal variables in the data frame. All character/factor columns
not included in this list will be considered as nominal variables, and
all numerical variables as numerical.}

\item{smote_K}{Relevant for SMOTE. The number of nearest neighbours for KNN.}

\item{verbose}{Whether to print progress to console.}

\item{random_seed}{Random seed for imbalance correction and cross-validation.}
}
\value{
A list with elements \code{train_set}, \code{test_set}, and \code{specs}, as returned
by \code{\link[=prep_to_train]{prep_to_train()}}. The object is also saved to disk as an .Rds file.
}
\description{
REMOVE THIS FUNCTION - IT IS UNUSED AND REDUNDANT !!!!!!!!
}
\details{
\code{prep_mm} Takes the path to the R object that resulted from the
pipeline ending with the \code{train_test_separation} function, prepares the
data for training using \code{prep_to_train} and trains using \code{train_mm}.
}
