% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_cols.R
\name{remove_duplicate_cols}
\alias{remove_duplicate_cols}
\title{Change duplicate entries to NA}
\usage{
remove_duplicate_cols(x)
}
\arguments{
\item{x}{A vector or array.}
}
\value{
The input vector with duplicated elements replaced by \code{NA}.
}
\description{
\code{remove_duplicate_cols} takes a vector or array and returns it with
duplicated elements replaced by NAs.
}
