% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns based on dictionary}
\usage{
rename_columns(
  df,
  colname_file = NULL,
  field_names = NULL,
  new_cols = NULL,
  ignore_cols = NULL
)
}
\arguments{
\item{df}{Input data frame whose column names will be changed.}

\item{colname_file}{Data frame that contains as separate columns both old
and new column names of \code{df}.}

\item{field_names}{Name of the column with UKB field names.}

\item{new_cols}{Name of the column with new column names}

\item{ignore_cols}{List of column names which should be ignored and their names
left unchanged.}
}
\value{
The input data frame with column names renamed according to the
provided dictionary or the default RAP-to-local conversion.
}
\description{
\code{rename_columns} takes a data frame and a .csv file that acts as a dictionary
to change column names of the data frame. It returns the data frame with
changed column names.
}
\details{
One of two options is allowed:
1: \code{field_names}, \code{new_cols}, and \code{colname_file_path} are \code{NULL}, in which case
the RAP-based column naming will be replae with the "old" local naming.
2: \code{field_names}, \code{new_cols}, and \code{colname_file_path} are provided by the user,
in which case the latter must contain as columns \code{field_names} and \code{new_cols}
that refer to the old to-be-replaced field names and the new names, respectively.
}
