% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_string.R
\name{table_to_string}
\alias{table_to_string}
\title{Create final features out of the variables}
\usage{
table_to_string(
  file_path,
  id_column_name,
  instances_low,
  instances_high,
  repeats_low,
  repeats_high,
  keep_assessments = "all"
)
}
\arguments{
\item{file_path}{File path of the table.}

\item{id_column_name}{The name of the column that contains field IDs.}

\item{instances_low}{Column name for the starting digit of instances.}

\item{instances_high}{Column name for the last digit of instances.}

\item{repeats_low}{Column name for the starting digit for repeats which are
0 or 1 for a0 and a1.}

\item{repeats_high}{Column name for the last digit for repeats.}

\item{keep_assessments}{Which assessments are to be kept; accepts 'all' and '0'.}
}
\value{
A single comma-separated character string of field ID names
formatted for the UKB table exporter.
}
\description{
\code{table_to_string} takes a table and transforms it into a string for the UKB
table exporter
}
