% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_mm_features.R
\name{amend_mm_features}
\alias{amend_mm_features}
\title{Correct class imbalance}
\usage{
amend_mm_features(df, mm_source = "self_report")
}
\arguments{
\item{df}{The input data frame.}

\item{mm_source}{The source of disease history.}
}
\description{
\code{amend_mm_features} takes the data frame and amends the MM features with
additional data.
}
