% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_disease_self.R
\name{create_diseases_self}
\alias{create_diseases_self}
\title{Create a data frame with multimorbidity columns}
\usage{
create_diseases_self(df, mm_codes, cancer = FALSE)
}
\arguments{
\item{df}{The input data frame.}

\item{mm_codes}{A data frame of disorder codes with the columns \code{code},
\code{disorder_long}, \code{field_id}, and \code{disorder}, indicating the self-report code
in UKB, the custom name of the disorder, the UKB field ID (20002 or 20001),
and a short name for the disorder, respectively.}

\item{cancer}{TRUE or FALSE; indicates whether field ID 20001 or 20002 is
used.}
}
\description{
\code{create_diseases_self} takes a data frame with an id column and the 20002 or
20001 UKB columns and returns a data frame with relevant disorders as rows
}
