% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mm_masterfile.R
\name{create_mm_masterfile}
\alias{create_mm_masterfile}
\title{Create the MM data frame}
\usage{
create_mm_masterfile(df, min_age, mm_source, mm_codes_file, random_seed = 24)
}
\arguments{
\item{df}{The input data frame.}

\item{min_age}{The minimum allowed age of participants.}

\item{mm_source}{The source of disease history.}

\item{mm_codes_file}{The path to the codes for each MM.}

\item{random_seed}{The number of \code{set.seed}.}
}
\description{
\code{create_mm_masterfile} takes the data frame with all the UKB data and
returns a data frame with participants as rows and binary indicators for
each disorder as columns.
}
