% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_indicator.R
\name{disease_indicator}
\alias{disease_indicator}
\title{Create a disease indicator from disease date}
\usage{
disease_indicator(
  df,
  colname,
  date_form = NULL,
  invalid_dates = NULL,
  threshold_date = NULL
)
}
\arguments{
\item{df}{Input data frame.}

\item{colname}{Character; Name of the column with the dates for the event.}

\item{date_form}{Character; format of the date.}

\item{invalid_dates}{Array; strings of invalid dates in the form \code{date_form}.
Events with these dates will be converted to NA.}

\item{threshold_date}{Single character or vector of same length as \code{nrow(df)};
should be in the same format as \code{colname}. If not \code{NULL}, events occurring
after this date are considered non-events and changed to NA in the returned
data frame.}
}
\value{
The input data frame with the original event column renamed to
\verb{<colname>_date} and a new integer indicator column \verb{<colname>} (1 = event,
0 = no event, \code{NA} = invalid date).
}
\description{
\code{disease_indicator} takes a data frame and a string indicating the column
containing the dates of an event, and returns the same data frame with
a new indicator column for that event. The original event column names is
changed; the new column has the same name as the original column for the
event.
}
