% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diagnoses.R
\name{extract_diagnoses}
\alias{extract_diagnoses}
\title{Create a long-format inpatient dataset}
\usage{
extract_diagnoses(
  df,
  raw = 1,
  colname_id = "eid",
  code_table = NULL,
  out_path = NULL,
  out_file_name = NULL
)
}
\arguments{
\item{df}{Input data frame with columns containing inpatient data. See details
for formatting requirements.}

\item{raw}{0 or 1; indicates whether the data is in its raw form - wide format
with each participant as a row - or already in the long format.}

\item{colname_id}{Character; the name of the column with participant IDs.}

\item{code_table}{Path to a .csv file or a data frame with two columns:
the first with diagnostic codes as strings and a second with the coding
format, also as string. See details.}

\item{out_path}{An optional path to the folder to which the output
data frame is to be written.}

\item{out_file_name}{The file name for the data frame if it is to be exported;
don't forget to include '.csv' at the end.}
}
\description{
\code{extract_diagnoses} takes a data frame with UKB data and returns a data frame
with select inpatient information.
}
\details{
If \code{raw = 1}, \code{df} must contain data on UKB field IDs 41270, 41271, 41280,
and 41281. The column names must be in the form X.ID.0.a, where X is the
letter "X", ID is the field ID, and a is the instance.
#' E.g., "X41270.0.1", "X41270.0.2",etc. Alternatively, if \code{raw = 0}, \code{df}
must contain 4 columns: participant ID column as character, diagnostic code as
character, date of diagnosis as character, and coding as character. The coding
indicates the coding system - ICD9, ICD10, etc. - for the diagnosis.

The allowed inputs for coding systems in the \code{code_table} data frame are
"icd9" and "icd10".
}
