% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_synthetic.R
\name{extract_from_synthetic}
\alias{extract_from_synthetic}
\title{Create a subset of the UKB synthetic data for your project}
\usage{
extract_from_synthetic(
  field_table,
  data_dict,
  id_column_name,
  data_path,
  out_path = NULL
)
}
\arguments{
\item{field_table}{.xlsx file. A table of IDs required for the project;
must contain a column with numerical field IDs as per the UKB showcase.}

\item{data_dict}{.csv file. A data-type lookup table that contains two columns:
\code{field_id} which is a numerical field ID as per the UKB showcase and
\code{type} which indicates the variable type for a given field - \code{datetime},
\code{int}, \code{real}, \code{string}, and \code{hes}; the latter refers to any fields that use
HES inpatient tables.'}

\item{id_column_name}{The name of the column in \code{field_table} that contains
the field IDs.}

\item{data_path}{The path for the .Rds files with synthetic UKB data.}

\item{out_path}{An optional path to write \code{masterfile_syn.csv} to, the finished
data frame that is returned by this function.}
}
\description{
\code{extract_from_synthetic} takes a table with field IDs, required .Rds files
with UKB synthetic data, and a data-type lookup table and returns a subset
of the synthetic data.
table exporter
}
