% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux_vars.R
\name{get_aux_vars}
\alias{get_aux_vars}
\title{Get auxiliary variables used for imputation and/or SMOTE}
\usage{
get_aux_vars(df)
}
\arguments{
\item{df}{The input data frame.}
}
\description{
\code{get_aux_vars} takes the data frame UKB participants and
returns a data frame with additional columns relevant for multimorbidity.
}
