% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_provider.R
\name{get_data_provider}
\alias{get_data_provider}
\title{Determine the data provider and censoring times for each participant}
\usage{
get_data_provider(
  df,
  hesin_path,
  gp_regs_path,
  gp_clinical_path,
  cens_dates,
  type,
  fill_NAs = NULL,
  random_seed
)
}
\arguments{
\item{df}{The input data frame.}

\item{hesin_path}{Path to the "hesin.txt" file, UKB category 2006.}

\item{gp_regs_path}{Path to GP registration records, UKB field ID 42038.}

\item{gp_clinical_path}{Path to GP clinical event records, UKB field ID 42040.}

\item{cens_dates}{A list of three dates in the form dd.mm.YYYY, corresponding}

\item{type}{Determines the approach used to assign the data provider. 'last'
assigns the last data provider that the participant was in contact with,
'freq' assigns the data provider that was the most frequent across the entire
period of electronic records.}

\item{fill_NAs}{For some participants - those that never went to either the
hospital or the GP - the data provider will not be found. This argument
determines how NAs should be handled. \code{NULL} leaves the NAs, 'random'
assigns a random data provider from the non-NAs, while 'HES', 'SMR', and
'PEDW' replace all NAs with the censoring date of the corresponding data provider.}

\item{random_seed}{Relevant only when \code{fill_NAs = 'random'}; the random seed
set before the sampling.}
}
\description{
\code{get_data_provider} Takes an input data frame with the UKB field id 40022
and 191, path names for registration and clinical event files, and a list
of dates to output a data frame with last data provider recorded in the record,
the most frequent data provider recorded in the record, and the censoring date.
}
