% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit_comparison.R
\name{matchit_comparison}
\alias{matchit_comparison}
\title{Compare performance of multiple distance-method combinations of matching}
\usage{
matchit_comparison(
  df,
  exposure,
  match_vars = setdiff(colnames(df), exposure),
  estimand,
  methods,
  distances,
  s.weights = NULL,
  imputed,
  verbose = FALSE,
  nnet_size = 100,
  nnet_MaxNWts = 1e+06,
  cardinality_solver = "highs"
)
}
\arguments{
\item{df}{Data frame or \code{mids} object containing the to-be-matched data.}

\item{exposure}{A string indicating the name of the column with the exposure
variable; must be binary.}

\item{match_vars}{A list of strings indicating the names of columns with
matching variables; by default, all non-exposure columns are used.}

\item{estimand}{A string for the estimand: 'ATE', 'ATT', or 'ATC'.}

\item{methods}{A list of methods as specified in \code{MatchIt}.}

\item{distances}{A list of distance metrics or algorithms for propensity score
calculation as specified in \code{MatchIt}.}

\item{s.weights}{Survey weights if appropriate.}

\item{imputed}{Logical; whether the input data have been imputed using MI or
not and thus whether \code{MatchThem} or \code{MatchIt} should be implemented.}

\item{verbose}{Logical; whether the method/distance currently evaluated
should be printed out.}

\item{nnet_size}{Size of the hidden layer for when \code{distance = nnet}.
See \code{nnet::nnet}.}

\item{nnet_MaxNWts}{Maximum number of allowed weights for when \code{distance = nnet}.
See \code{nnet::nnet}.}
}
\description{
\code{matchit_comparison} returns a list, with different matching approaches as
rows and performance metrics as columns. Implements \code{MatchIt} and \code{MatchtThem}.
duplicated elements replaced by NAs.
}
