% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit_extract.R
\name{matchit_extract}
\alias{matchit_extract}
\title{Extract simplified table of matching metrics from output of \code{matchit_comparison}.}
\usage{
matchit_extract(bal, imputed, smd_thresh = 0.1)
}
\arguments{
\item{bal}{The input list; should be the output of \code{matchit_comparison}.}

\item{smd_thresh}{The SMD threshold for what is considered a good match.}

\item{impute}{Whether the matching was performed on an imputed dataset.}
}
\description{
\code{matchit_extract} takes as input a list created from running the
\code{matchit_comparison} function and returns a data frame with simplified
matching metrics, with matching approaches as rows and metrics as columns.
}
