% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_to_train.R
\name{prep_to_train}
\alias{prep_to_train}
\title{Preparation of the data for training and testing. Inludes potential feature
amendments, data filtering, and imbalance correction.}
\usage{
prep_to_train(
  train_set,
  test_set,
  target_var,
  amend_features = FALSE,
  max_followup = NULL,
  remove_censored = FALSE,
  imbalance_correct = NULL,
  remove_vars = NULL,
  min_age = NULL,
  balance_prop,
  random_seed,
  ordinals,
  smote_K,
  verbose = verbose
)
}
\arguments{
\item{train_set}{The training set.}

\item{test_set}{The test set.}

\item{target_var}{The target variable to be predicted.}

\item{amend_features}{Whether features should be amended; relevant only for
self-report data.}

\item{max_followup}{The time within which the outcome that is to be
predicted should occur.}

\item{remove_censored}{Whether participants that were lost to follow-up
before max_followup should be removed. Assumes the presence of the columns
\code{asc_date} and \code{censor_date} in the data frame.}

\item{imbalance_correct}{Whether imbalance correction should be performed.
Either \code{NULL}, \code{'SMOTE'} or \code{'downsample'}.}

\item{remove_vars}{A list of column names that should be removed prior to
balance correction. Especially useful for SMOTE.}

\item{min_age}{The minimum age of the sample. Determines whether participants
below a certain age should be dropped prior to training.}

\item{balance_prop}{Only relevant if \code{imbalance_correct = 'SMOTE'}. The
proportion of observations that will belong to the erstwhile minority class
after SMOTE.}

\item{random_seed}{The random seed for SMOTE and downsampling.}

\item{ordinals}{Only relevant if \code{imbalance_correct = 'SMOTE'}. A vector of
ordinal categorical variables in the training data.}

\item{verbose}{Whether to print progress to console.}

\item{K}{Only relevant if \code{imbalance_correct = 'SMOTE'}. The K for KNN.}
}
\description{
\code{prep_to_train} Takes the train and test sets and returns them ready to be
input into a predictive model. All numerical variables are scaled.
}
