% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tbl.R
\name{read_tbl}
\alias{read_tbl}
\title{Reads in Excel table/sheet; detects format based on file suffix.}
\usage{
read_tbl(tbl_path)
}
\arguments{
\item{tbl_path}{Path to the table.}
}
\description{
\code{read_tbl} takes as input the path to an \verb{.xlsx,} \code{.xls}, or \code{.csv} table and
returns as the output the same table as an R dataframe.
}
