% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove numerical outliers}
\usage{
remove_outliers(x, method, var_metric)
}
\arguments{
\item{x}{Input array of numerical values.}

\item{method}{Character; indicates which statistic is used to calculate the
spread. Must be either 'sd' or 'iqr'.}

\item{var_metric}{Numerical; indicates the minimum number of variance units -
standard deviations or interquartile ranges - from the mean/median that are
considered outliers.}
}
\description{
\code{remove_outliers} takes an array, sets outliers to NA, and returns the array.
}
