% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns based on dictionary}
\usage{
rename_columns(df, colname_file, old_cols, new_cols)
}
\arguments{
\item{df}{Input data frame whose column names will be changed.}

\item{colname_file}{Data frame that contains as separate columns both old
and new column names of \code{df}.}

\item{old_cols}{Name of the column with old column names.}

\item{new_cols}{Name of the column with new column names}
}
\description{
\code{rename_columns} takes a data frame and a .csv file that acts as a dictionary
to change column names of the data frame.
}
