% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_birth_date.R
\name{set_birth_date}
\alias{set_birth_date}
\title{Sets birth date.}
\usage{
set_birth_date(df, yob, mob, day = "random", random_seed = 246)
}
\arguments{
\item{df}{Data frame.}

\item{yob}{Year of birth column name.}

\item{mob}{Month of birth column name.}

\item{day}{Whether the same day for all participants or a random day for each
participant is set. Set to \code{random} by default; otherwise a number between 1
and 28 can be input.}

\item{random_seed}{Random seed for selecting the day.}
}
\description{
\code{set_birth_date} takes as input the data frame and the names of columns
representing year of birth and month of birth, respectively. It returns
the input data frame with a new column for date of birth
}
