% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_string.R
\name{table_to_string}
\alias{table_to_string}
\title{Create final features out of the variables}
\usage{
table_to_string(file_path, id_column_name, instance_column_name, output_path)
}
\arguments{
\item{file_path}{File path of the table.}

\item{id_column_name}{The name of the column that contains field IDs.}

\item{instance_column_name}{The name of the column that contains the number
of instances.}

\item{output_path}{The path, including full file name, where to write the file.}
}
\description{
\code{table_to_string} takes a table and transforms it into a string for the UKB
table exporter
}
