% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_mm.R
\name{train_mm}
\alias{train_mm}
\title{Train the prediction algorithm, potentially correcting for class imbalance.}
\usage{
train_mm(
  df,
  target_var,
  features,
  amend_features = FALSE,
  min_age,
  max_followup = NULL,
  algorithm,
  train_metric = "ROC",
  cv_folds,
  tune_len,
  distribute_cores,
  core_number,
  random_seed,
  verbose = FALSE,
  imbalance_correct = NULL,
  balance_prop = NULL,
  ordinals = NULL,
  smote_K = NULL
)
}
\arguments{
\item{df}{The input data frame.}

\item{target_var}{The target variable to be predicted.}

\item{amend_features}{Whether features should be amended; relevant only for
self-report data.}

\item{min_age}{The minimum age of the sample. Determines whether participants
below a certain age should be dropped prior to training.}

\item{max_followup}{The time within which the outcome that is to be
predicted should occur. If not set to \code{NULL}, the input data frame must
contain a column called 'followup' which is the time in years between baseline
and right-censoring or event occurrence, whichever occurs first.}

\item{algorithm}{The algorithm to be used for prediction.}

\item{train_metric}{Which metric to optimise on. Available options are
"Accuracy", "Kappa", "ROC", "Sens", and "Spec".}

\item{cv_folds}{The number of folds to be used during cross-validation;
corresponds to the argument \code{number} in \code{caret::trainControl}.}

\item{tune_len}{The length of the tuning grid; corresponds to the
\code{tuneLength} argument of \code{caret::train}.}

\item{distribute_cores}{Whether cross-validation should be performed in
parallel.}

\item{core_number}{Relevant only if \code{distribute_cores = TRUE}. The number of
cores to be used in parallel processing.
#' @param random_seed Random seed used in 1. balance adjustment, 2. registering
doRNG for reproducible parallel processing, and 3. generation of further
seeds to be used in the generation of the cross-validation folds and of the
tuning grid - argument \code{seeds} in \code{caret::trainControl}.}

\item{verbose}{Whether to print progress to console.}

\item{imbalance_correct}{Whether potential class imbalance between levels
of the target variables should be corrected.}

\item{balance_prop}{Relevant only in the case of imbalance correction; the desired
proportion of the total observations for the current minority class.}

\item{ordinals}{Relevant only when \code{imbalance_correct = 'SMOTE'}; a list of all
ordinal variables in the data frame.}

\item{smote_K}{Relevant only when \code{imbalance_correct = 'SMOTE'}; determines the
K used in KNN during SMOTE.}
}
\description{
\code{train_mm} Takes and input data frame, potentially corrects for class
imbalance of the target variable, and trains a predictive model using
the supplied features.
}
