% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_mm.R
\name{train_mm}
\alias{train_mm}
\title{Train the prediction algorithm, potentially correcting for class imbalance.}
\usage{
train_mm(
  df,
  target_var,
  features,
  algorithm,
  train_metric = "ROC",
  cv_folds,
  tune_len,
  distribute_cores,
  core_number,
  random_seed,
  verbose = FALSE
)
}
\arguments{
\item{df}{The input data frame.}

\item{target_var}{The target variable to be predicted.}

\item{algorithm}{The algorithm to be used for prediction.}

\item{train_metric}{Which metric to optimise on. Available options are
"Accuracy", "Kappa", "ROC", "Sens", and "Spec".}

\item{cv_folds}{The number of folds to be used during cross-validation;
corresponds to the argument \code{number} in \code{caret::trainControl}.}

\item{tune_len}{The length of the tuning grid; corresponds to the
\code{tuneLength} argument of \code{caret::train}.}

\item{distribute_cores}{Whether cross-validation should be performed in
parallel.}

\item{core_number}{Relevant only if \code{distribute_cores = TRUE}. The number of
cores to be used in parallel processing.
#' @param random_seed Random seed used in 1. balance adjustment, 2. registering
doRNG for reproducible parallel processing, and 3. generation of further
seeds to be used in the generation of the cross-validation folds and of the
tuning grid - argument \code{seeds} in \code{caret::trainControl}.}

\item{verbose}{Whether to print progress to console.}
}
\description{
\code{train_mm} Takes and input data frame, potentially corrects for class
imbalance of the target variable, and trains a predictive model using
the supplied features.
}
