% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_separation.R
\name{train_test_separation}
\alias{train_test_separation}
\title{Separate the data into train and test sets}
\usage{
train_test_separation(
  df,
  train_prop,
  random_seed,
  impute_missing,
  exclude_vars = c("id"),
  imp_threads = 1,
  max_iter = 10,
  num_trees = 500,
  pmm_k = 5,
  imp_seed,
  verbose = verbose
)
}
\arguments{
\item{df}{The input data frame.}

\item{train_prop}{Proportion of observations to be allocated to train set.}

\item{random_seed}{Random seed set just prior to train/test split.}

\item{impute_missing}{Whether missing observations should be imputed with
Random Forest using \code{missRanger::missRanger}. It assumes that there is a
column called \code{id} that will not be used in the imputation; all other
columns of the data frame are used.}

\item{exclude_vars}{Whether any variables should be excluded from the
imputation. By default, the column \code{id} is excluded; if all should be used,
set to \code{NULL}.}

\item{imp_threads}{How many threads should be used for the imputation.}

\item{max_iter}{Number of iterations during imputation; corresponds to
\code{max.iter} argument in \code{missRanger} and relevant only when
\code{impute_missing = TRUE}.}

\item{num_trees}{Number of trees to build during imputation; corresponds to
\code{num.trees} argument in \code{missRanger} and relevant only when
\code{impute_missing = TRUE}.}

\item{pmm_k}{Number of values to sample from in predictive mean matching after
imputation; corresponds to \code{pmm.k} argument in \code{missRanger} and relevant only when
\code{impute_missing = TRUE}.}

\item{imp_seed}{Seed for the imputation; corresponds to \code{seed} argument in
\code{missRanger} and relevant only when \code{impute_missing = TRUE}.}

\item{verbose}{Whether to print progress to console.}
}
\description{
\code{train_test_separation} takes a data frame with all features required for
training and separates it into train and test sets.
}
