% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_separation.R
\name{train_test_separation}
\alias{train_test_separation}
\title{Separate the data into train and test sets}
\usage{
train_test_separation(
  df,
  target,
  train_prop,
  random_seed,
  impute_missing,
  imp_threads = 1,
  max_iter = 10,
  num_trees = 500,
  pmm_k = 5,
  imp_seed
)
}
\arguments{
\item{df}{The input data frame.}

\item{target}{The name of the column with the target variable.}

\item{train_prop}{Proportion of observations to be allocated to train set.}

\item{random_seed}{Random seed set just prior to train/test split.}

\item{impute_missing}{Whether missing observations should be imputed with
Random Forest using \code{missRanger::missRanger}.}

\item{imp_threads}{How many threads should be used for the imputation.}

\item{max_iter}{Number of iterations during imputation; corresponds to
\code{max.iter} argument in \code{missRanger} and relevant only when
\code{impute_missing = TRUE}.}

\item{num_trees}{Number of trees to build during imputation; corresponds to
\code{num.trees} argument in \code{missRanger} and relevant only when
\code{impute_missing = TRUE}.}

\item{pmm_k}{Number of values to sample from in predictive mean matching after
imputation; corresponds to \code{pmm.k} argument in \code{missRanger} and relevant only when
\code{impute_missing = TRUE}.}

\item{imp_seed}{Seed for the imputation; corresponds to \code{seed} argument in
\code{missRanger} and relevant only when \code{impute_missing = TRUE}.}
}
\description{
\code{train_test_separation} takes a data frame with all variables required for
training and separates it into train and test sets. The target variable -
if present in the data frame - will be removed.
}
