% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_reclassify.R
\name{ukb_reclassify}
\alias{ukb_reclassify}
\title{Reclassify various UKB variables}
\usage{
ukb_reclassify(x, field_id)
}
\arguments{
\item{x}{The input array.}

\item{field_id}{Integer; field ID for the variable.}
}
\description{
\code{ukb_reclassify} takes an array of coded observations for various UKB
variables and re-classifies them. See \code{details} for more.
qualifications answers
}
\details{
6138 (qualifications): classification into three classes: higher,
secondary/vocational, none. The higher activity level is prioritised in
cases of answers fitting into multiple categories per participant.

6164 (physical activity): classification into four classes:
0 - none (-7, none of the above),
1 - low (code 4: light household tasks only),
2 - medium (codes 1/2/5: heavy household tasks, walking for pleasure, other exercise),
3 - high (code 3: strenuous sports).
}
