% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKBfunctions-package.R
\docType{package}
\name{UKBfunctions-package}
\alias{UKBfunctions}
\alias{UKBfunctions-package}
\title{UKBfunctions: Utilities for Processing and Analysing UK Biobank Data}
\description{
Provides utilities for processing and analysing UK Biobank (UKB) data for multimorbidity research and predictive modelling. Includes functions for disease ascertainment from self-report and inpatient data, feature engineering of auxiliary variables, stratified train/test splitting with optional imputation, class imbalance correction, and model training via 'caret'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JuM24/UKBfunctions}
  \item Report bugs at \url{https://github.com/JuM24/UKBfunctions/issues}
}

}
\author{
\strong{Maintainer}: Jure Mur \email{jure.mur@uni.lu}

Authors:
\itemize{
  \item Jure Mur \email{jure.mur@uni.lu}
}

}
\keyword{internal}
