% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_dummies.R
\name{undo_dummies}
\alias{undo_dummies}
\title{Undo hot-one-encoding}
\usage{
undo_dummies(df, remove_dummies = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{The input data frame.}

\item{remove_dummies}{Whether the dummy variables should be removed.}

\item{verbose}{Whether to print progress to console.}
}
\description{
\code{undo_dummies} takes a data frame with dummy variables created by
hot-one-encoding, often through the use of KNN-based methods, and
converts the dummy variables into their original multi-level form.
}
